+++ app/lib/datamodelmodeler/features/grid-snapping/index.js
@@ -1,4 +1,4 @@
-import BpmnGridSnapping from './BpmnGridSnapping';
+import BpmnGridSnapping from '../../../frequentlyUsedFiles/features/grid-snapping/BpmnGridSnapping';
 import GridSnappingModule from 'diagram-js/lib/features/grid-snapping';
 
 import GridSnappingBehaviorModule from './behavior';
