+++ app/lib/rolemodeler/features/di-ordering/index.js
@@ -1,4 +1,4 @@
-import ODDiOrdering from './ODDiOrdering';
+import ODDiOrdering from '../../../frequentlyUsedFiles/features/di-ordering/ODDiOrdering';
 
 export default {
   __init__: [
